requirements.txt
Flask
PyQt5
python-chess
tensorflow
numpy
matplotlib
tqdm
google-cloud-storage
torch
torchvision
gunicorn
flask-cors==5.0.0